Fetch links and profile data in HomePageContent
