+++ src/app/components/Links.tsx
@@ -5,7 +5,6 @@ import Button from '../common/components/Button'
 import LinksList from './LinksList'
 import { useMutation, useQuery } from '@tanstack/react-query'
 import { addLink, editLinks, getLinks } from '../services/links'
-import { useEffect } from 'react'
 import { useAppDispatch } from '../common/hooks/useAppDispatch'
 import { ModifiedLink, setList } from '@/lib/store/slices/linksSlice'
 import { toast } from 'react-toastify'
@@ -34,14 +33,6 @@ const Links = () => {
   const links = useAppSelector((state) => state.links.list);
 
   const modifiedLinks = links?.filter((link) => link.isModified);
-
-  useEffect(() => {
-    if (!getLinksQuery.data) {
-      return;
-    }
-
-    dispatch(setList(getLinksQuery.data));
-  }, [getLinksQuery.data, dispatch])
 
   const addNewLink = async () => {
     try {
